Feat/userid auth check

+++ src/main/java/com/example/monimentoom/domain/position/service/PositionService.java
@@ -10,9 +10,9 @@ import com.example.monimentoom.domain.room.model.Room;
 import com.example.monimentoom.domain.room.repository.RoomRepository;
 import com.example.monimentoom.exception.CustomException;
 import com.example.monimentoom.exception.ErrorCode;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -25,19 +25,15 @@ public class PositionService {
      * 굿즈 새로 배치
      */
     @Transactional
-    public PositionResponse createPosition(PositionRequest request) {
-        // TODO : 추후 해당 사용자가 맞는 지 임시 검토 로직,
-//        User user = userRepository.findById(userId)
-//                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
+    public PositionResponse createPosition(Long userId, PositionRequest request) {
         Room room = roomRepository.findById(request.getRoomId())
                 .orElseThrow(() -> new CustomException(ErrorCode.ROOM_NOT_FOUND));
 
-//        if (!room.getUser().getId().equals(userId)) {
-//            throw new IllegalArgumentException("이 방을 수정할 권한이 없습니다.");
-//        }
+        room.validateOwnership(userId);
 
         Goods goods = goodsRepository.findById(request.getGoodsId())
                 .orElseThrow(() -> new CustomException(ErrorCode.GOODS_NOT_FOUND));
+        goods.validateOwnership(userId);
         Position position = Position.builder()
                 .room(room)
                 .goods(goods)
@@ -52,17 +48,23 @@ public class PositionService {
     }
 
     @Transactional
-    public PositionResponse updatePosition(Long positionId, PositionRequest request) {
-//        User user = userRepository.findById(userId)
-//                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
-
+    public PositionResponse updatePosition(Long userId, Long positionId, PositionRequest request) {
         Position position = positionRepository.findById(positionId)
                 .orElseThrow(() -> new CustomException(ErrorCode.POSITION_NOT_FOUND));
-        Room room = roomRepository.findById(request.getRoomId())
-                .orElseThrow(() -> new CustomException(ErrorCode.ROOM_NOT_FOUND));
+        // 기존 방도 사용자 것이 맞는지 검증
+        Room currentRoom = position.getRoom();
+        currentRoom.validateOwnership(userId);
 
-        position.update(room, request.getX(), request.getY(), request.getWallSide(), request.getWidthUnit(), request.getHeightUnit());
-
+        // 방 이동이 없다면, DB 조회 생략하여 성능 향상
+        Room targetRoom;
+        if (currentRoom.getId().equals(request.getRoomId())) {
+            targetRoom = currentRoom;
+        } else {
+            targetRoom = roomRepository.findById(request.getRoomId())
+                    .orElseThrow(() -> new CustomException(ErrorCode.ROOM_NOT_FOUND));
+            targetRoom.validateOwnership(userId);
+        }
+        position.update(targetRoom, request.getX(), request.getY(), request.getWallSide(), request.getWidthUnit(), request.getHeightUnit());
         // update이므로 Transactional로 save 대체 하였음.
         return PositionResponse.from(position);
     }
@@ -71,9 +73,11 @@ public class PositionService {
      * 굿즈 배치 해제
      */
     @Transactional
-    public void deletePosition(Long positionId) {
+    public void deletePosition(Long userId, Long positionId) {
         Position position = positionRepository.findById(positionId)
                 .orElseThrow(() -> new CustomException(ErrorCode.POSITION_NOT_FOUND));
+        Room room = position.getRoom();
+        room.validateOwnership(userId);
 
         positionRepository.delete(position);
     }
